Split `Error` into two
