+++ test-lib/src/firmware.rs
@@ -1,7 +1,12 @@
+pub mod error;
 pub mod receive;
 pub mod send;
 
 pub use self::{
+    error::{
+        Error,
+        Result,
+    },
     receive::Receiver,
     send::Sender,
 };
