+++ test-lib/src/host.rs
@@ -8,8 +8,6 @@ use serde::{
     Serialize,
 };
 
-use crate::Result;
-
 
 /// Send a message to the target, via the provided writer
 ///
@@ -51,4 +49,31 @@ pub fn receive<'de, T, R>(mut reader: R, buf: &'de mut Vec<u8>) -> Result<T>
 
     let event = postcard::from_bytes_cobs(buf)?;
     Ok(event)
+}
+
+
+pub type Result<T = ()> = core::result::Result<T, Error>;
+
+
+#[derive(Debug)]
+pub enum Error {
+    /// An I/O error occured
+    Io(io::Error),
+
+    /// An error originated from Postcard
+    ///
+    /// The `postcard` crate is used for (de-)serialization.
+    Postcard(postcard::Error),
+}
+
+impl From<io::Error> for Error {
+    fn from(err: io::Error) -> Self {
+        Self::Io(err)
+    }
+}
+
+impl From<postcard::Error> for Error {
+    fn from(err: postcard::Error) -> Self {
+        Self::Postcard(err)
+    }
 }
